se.bluebrim.crud.client/src/test/java/se/bluebrim/common/client/command/CommandChainTest.java
package se.bluebrim.common.client.command;

import static org.junit.Assert.assertTrue;

import java.util.concurrent.CountDownLatch;
import java.util.concurrent.TimeUnit;

import org.junit.BeforeClass;
import org.junit.Test;

import se.bluebrim.crud.client.command.BackgroundCommand;
import se.bluebrim.crud.client.command.CommandChain;
import se.bluebrim.crud.client.command.CursorHandler;
import se.bluebrim.crud.client.command.DefaultBackgroundCommand;
import se.bluebrim.crud.client.command.DefaultCommand;
import se.bluebrim.crud.client.command.ExceptionHandler;
import se.bluebrim.crud.client.command.ProgressVisualizer;

public class CommandChainTest
{
	
	@BeforeClass
	public static void setUpBeforeClass() throws Exception
	{
		DefaultCommand.setDefaultCursorHandler(new CursorHandler(){

			@Override
			public void setWaitCursor()
			{
				System.out.println("set Wait Cursor");				
			}
		
			@Override
			public void resetWaitCursor()
			{
				System.out.println("reset Wait Cursor");				
			}}

		);
		
		DefaultCommand.setDefaultProgressVisualizer(new ProgressVisualizer(){

			@Override
			public void startProgressAnimation()
			{
				System.out.println("start Progress Animation");				
			}

			@Override
			public void stopProgressAnimation()
			{
				System.out.println("stop Progress Animation");				
			}});
		
		DefaultCommand.setDefaultExceptionHandler(new ExceptionHandler(){

			@Override
			public void handleException(Throwable t)
			{
				System.out.println("Exception: " + t);			
			}});
	}
	
	/**
	 * The test environment thinks this is failure. Exclude this test until
	 * we can figure out how to test this.
	 */
	public void testFirstFails() throws Exception
	{
		final CountDownLatch latch = new CountDownLatch(1);		
		new CommandChain(new FailingCommand("Command 1"), new NormalCommand("Command 2")).run(new BackgroundCommand.DoneListener(){

			@Override
			public void done()
			{
				latch.countDown();
				
			}});
		assertTrue(latch.await(10, TimeUnit.SECONDS));
	}
		
	@Test
	public void testTwoNormal() throws Exception
	{
		final CountDownLatch latch = new CountDownLatch(1);		
		new CommandChain(new NormalCommand("Command 1"), new NormalCommand("Command 2")).run(new BackgroundCommand.DoneListener(){

			@Override
			public void done()
			{
				latch.countDown();
				
			}});
		assertTrue(latch.await(10, TimeUnit.SECONDS));
	}
	
	@Test
	public void testThreeNormal() throws Exception
	{
		final CountDownLatch latch = new CountDownLatch(1);		
		new CommandChain(new NormalCommand("Command 1"), new NormalCommand("Command 2"), new NormalCommand("Command 3 and final")).run(new BackgroundCommand.DoneListener(){

			@Override
			public void done()
			{
				latch.countDown();
				
			}});
		assertTrue(latch.await(10, TimeUnit.SECONDS));
	}
		
	@Test
	public void testFirtsCancels() throws Exception
	{
		final CountDownLatch latch = new CountDownLatch(1);		
		new CommandChain(new CancelingCommand("Command 1"), new NormalCommand("Command 2")).run(new BackgroundCommand.DoneListener(){

			@Override
			public void done()
			{
				latch.countDown();
				
			}});
		assertTrue(latch.await(10, TimeUnit.SECONDS));
	}

			
	private class FailingCommand extends DefaultBackgroundCommand
	{
		private String name;
				
		public FailingCommand(String name)
		{
			super();
			this.name = name;
		}

		@Override
		protected Object runInBackground() throws Exception
		{
			System.out.println("Start in background: " + name);
			Thread.sleep(2000);
			throw new RuntimeException("Oops!");
		}
		
		@Override
		protected void publishResult(Object result) throws Exception
		{
			System.out.println("Publish result from: " + result);			
		}
		
	}
	
	private class NormalCommand extends DefaultBackgroundCommand
	{
		private String name;
				
		public NormalCommand(String name)
		{
			super();
			this.name = name;
		}

		@Override
		protected Object runInBackground() throws Exception
		{
			System.out.println("Start in background: " + name);
			Thread.sleep(1000);
			return name;
		}
		
		@Override
		protected void publishResult(Object result) throws Exception
		{
			System.out.println("Publish result from: " + result);			
		}
		
	}
	
	private class CancelingCommand extends NormalCommand
	{

		public CancelingCommand(String name)
		{
			super(name);
		}
		
		@Override
		protected void prepare()
		{
			throw new DefaultBackgroundCommand.CancelCommandException("Sorry must cancel now");
		}
		
	}

		


}
